Interfaces Refactor
