--- src/interfaces/ProductCategoriesResponse.ts
@@ -1,58 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface ProductCategoriesState {
-  data: ProductCategoriesResponse;
-  isLoading: boolean;
-}
-
-export interface ProductCategoriesResponse {
-  page: number;
-  results_per_page: number;
-  results_size: number;
-  total_results_size: number;
-  total_pages: number;
-  next_page: null;
-  prev_page: null;
-  results: Result[];
-  version: string;
-  license: string;
-}
-
-export interface Result {
-  id: string;
-  uid: null;
-  url: null;
-  type: string;
-  href: string;
-  tags: any[];
-  first_publication_date: string;
-  last_publication_date: string;
-  slugs: string[];
-  linked_documents: any[];
-  lang: string;
-  alternate_languages: AlternateLanguage[];
-  data: Data;
-}
-
-export interface AlternateLanguage {
-  id: string;
-  type: string;
-  lang: string;
-}
-
-export interface Data {
-  name: string;
-  main_image: MainImage;
-}
-
-export interface MainImage {
-  dimensions: Dimensions;
-  alt: string;
-  copyright: null;
-  url: string;
-}
-
-export interface Dimensions {
-  width: number;
-  height: number;
-}
