--- src/interfaces/ProductsResponse.ts
@@ -1,93 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface ProductsState {
-  data: ProductsResponse;
-  isLoading: boolean;
-}
-
-export interface ProductsResponse {
-  page: number;
-  results_per_page: number;
-  results_size: number;
-  total_results_size: number;
-  total_pages: number;
-  next_page: null;
-  prev_page: null;
-  results: Result[];
-  version: string;
-  license: string;
-}
-
-export interface Result {
-  id: string;
-  uid: null;
-  url: null;
-  type: string;
-  href: string;
-  tags: string[];
-  first_publication_date: string;
-  last_publication_date: string;
-  slugs: string[];
-  linked_documents: any[];
-  lang: string;
-  alternate_languages: AlternateLanguage[];
-  data: Data;
-}
-
-export interface AlternateLanguage {
-  id: string;
-  type: string;
-  lang: string;
-}
-
-export interface Data {
-  name: string;
-  sku: string;
-  category: Category;
-  mainimage: MainimageClass;
-  short_description: string;
-  description: Description[];
-  specs: Spec[];
-  images: ImageElement[];
-  stock: number;
-  price: number;
-}
-
-export interface Category {
-  id: string;
-  type: string;
-  tags: any[];
-  slug: string;
-  lang: string;
-  first_publication_date: string;
-  last_publication_date: string;
-  link_type: string;
-  isBroken: boolean;
-}
-
-export interface Description {
-  type: string;
-  text: string;
-  spans: any[];
-}
-
-export interface ImageElement {
-  image: MainimageClass;
-}
-
-export interface MainimageClass {
-  dimensions: Dimensions;
-  alt: null | string;
-  copyright: null;
-  url: string;
-}
-
-export interface Dimensions {
-  width: number;
-  height: number;
-}
-
-export interface Spec {
-  spec_name: string;
-  spec_value: string;
-}
