+++ src/interfaces/shared/AlternateLanguage.ts
@@ -1,0 +1,5 @@
+export default interface AlternateLanguage {
+  id: string;
+  type: string;
+  lang: string;
+}
