+++ src/interfaces/shared/Dimensions.ts
@@ -1,0 +1,4 @@
+export default interface Dimensions {
+  width: number;
+  height: number;
+}
