+++ src/interfaces/shared/MainImage.ts
@@ -1,0 +1,8 @@
+import Dimensions from './Dimensions';
+
+export default interface MainImage {
+  dimensions: Dimensions;
+  alt: string;
+  copyright: null;
+  url: string;
+}
