+++ src/interfaces/shared/index.ts
@@ -1,0 +1,5 @@
+import AlternateLanguage from './AlternateLanguage';
+import Dimensions from './Dimensions';
+import MainImage from './MainImage';
+
+export type { AlternateLanguage, Dimensions, MainImage };
